+++ cynic/src/result.rs
@@ -1,17 +1,3 @@
-/// The result of a GraphQL Operation.
-///
-/// Either it fully succeeds, or we have some errors.  If we have some errors
-/// then some fields might unexpectedly be null, so deserialization _might_ have
-/// failed.  We represent that with a PossiblyParsedData<T>.
-pub type GraphQlResult<T> = Result<T, (PossiblyParsedData<T>, Vec<GraphQlError>)>;
-
-#[deprecated(
-    since = "0.13.0",
-    note = "GraphQLResult has been deprecated in favour of GraphQlResult"
-)]
-#[allow(clippy::upper_case_acronyms)]
-pub type GraphQLResult<T> = GraphQlResult<T>;
-
 #[derive(Debug, serde::Deserialize)]
 pub struct GraphQlResponse<T> {
     pub data: Option<T>,
@@ -23,24 +9,8 @@ pub struct GraphQlResponse<T> {
     note = "GraphQLResponse has been deprecated in favour of GraphQlResponse"
 )]
 #[allow(clippy::upper_case_acronyms)]
+#[doc(hidden)]
 pub type GraphQLResponse<T> = GraphQlResponse<T>;
-
-/*
-impl<T> GraphQLResponse<T> {
-    fn into_result(self) -> GraphQLResult<T> {
-        // TODO: I need access to the query/selection set in order to decode this.
-        if let Some(errors) = self.errors {
-            if !errors.is_empty() {
-                // TODO: Return an error.
-            }
-        }
-
-        match self {
-            GraphQLResponse::Ok(t) => Ok(t),
-            GraphQLResponse::Err(poss_data, errs) => Err((poss_data, errs)),
-        }
-    }
-}*/
 
 /// A model describing an error which has taken place during execution.
 #[derive(Clone, Debug, Eq, PartialEq, serde::Deserialize, thiserror::Error)]
@@ -112,14 +82,3 @@ pub enum GraphQlErrorPathSegment {
 )]
 #[allow(clippy::upper_case_acronyms, dead_code)]
 pub type GraphQLErrorPathSegment = GraphQlErrorPathSegment;
-
-/// The data returned by a GraphQL query when the query had errors.
-/// GraphQL allows servers to return partial data in this case, but if there's
-/// missing fields that aren't represented by an Option we won't have been
-/// able to decode that data.
-pub enum PossiblyParsedData<T> {
-    ParsedData(T),
-    // TODO: Could pass serde_json::Value from here, can't be bothered right
-    // now though....
-    UnparsedData,
-}